changelog.txt 2017/1/24 v0.12, immediate-run-token, stacktrace

+++ Copper/src/utilList.h
@@ -9,7 +9,7 @@ namespace util {
 class IndexOutOfBoundsException {};
 class NullListNodeException {};
 
-typedef unsigned int	uint;
+typedef unsigned long	uint;
 
 template<class T>
 class List
